Disable audit certificates for new audit enrollments.
An `eligible_for_certificate` field is added to the
GeneratedCertificate model. This way we can retain existing grading
logic, as well as maintaining correctness in analytics and reporting.

Ineligible certs are hidden be default. This is done by customizing
the manager that GeneratedCertificate uses to filter out ineligible
certs by default. There are some places in the code (largely
reporting, analytics, and management commands) that use the fallback
option without filtering,
`GeneratedCertificate.with_ineligible_certificates`.